requirements.txt
openai
PyPDF2
fpdf
tiktoken
python-dotenv